[Feat] 포스트 리스트 컴포넌트 완성
- 기본 구조 변경
- aria-current 타입 오류 해결
- SSG + redux
- 태그(카테고리) 변경
- 페이지네이션 렌더링만 완성
