+++ src/components/postsList.module.scss
@@ -27,35 +27,6 @@
     }
   }
 
-  .category {
-    display: flex;
-    padding-top: 15px;
-    padding-bottom: 50px;
-
-    li {
-      height: 35px;
-      padding: 2px 15px;
-      margin-right: 15px;
-      color: colors.$BLACK;
-      word-break: keep-all;
-      cursor: pointer;
-      background-color: rgb(226, 232, 240);
-      border: none;
-      border-radius: 3px;
-      transition: 0.2s ease-out;
-
-      a {
-        width: 100%;
-        height: 100%;
-      }
-
-      &:hover {
-        color: colors.$WHITE;
-        background-color: colors.$PUPLE;
-      }
-    }
-  }
-
   .postList {
     display: grid;
     grid-template-rows: repeat(2, 1fr);
